refactor: split functions in utilities file
