+++ src/components/utils/__tests__/addCommas.spec.ts
@@ -1,0 +1,8 @@
+import { addCommas } from '../addCommas';
+
+describe('addCommas', () => {
+  it('should remove commas in string', () => {
+    const value = addCommas('1000000');
+    expect(value).toEqual('1,000,000');
+  });
+});
