+++ src/components/utils/__tests__/removeCommas.spec.ts
@@ -1,0 +1,8 @@
+import { removeCommas } from '../removeCommas';
+
+describe('removeCommas', () => {
+  it('should remove commas in string', () => {
+    const value = removeCommas('1,000,000');
+    expect(value).toEqual('1000000');
+  });
+});
