+++ src/components/utils/addCommas.ts
@@ -1,0 +1,1 @@
+export const addCommas = (value: string): string => value.replace(/\B(?=(\d{3})+(?!\d))/g, ',');
