+++ src/components/utils/removeCommas.ts
@@ -1,0 +1,1 @@
+export const removeCommas = (value: string): string => value.replace(/,/g, '');
